Renamed file GOOrganDialog to GOOrganSettingsDialog
